サンプルコードモジュール名、パッケージ名変更
