+++ sdk-example-kt/build.gradle
@@ -1,32 +1,31 @@
 apply plugin: 'com.android.application'
 
-android {
-    compileSdkVersion 28
+apply plugin: 'kotlin-android'
 
+apply plugin: 'kotlin-android-extensions'
+
+android {
+    compileSdkVersion versions.compileSdk
     defaultConfig {
-        applicationId "jp.studyplus.android.sdk_example_java"
-        minSdkVersion 16
-        targetSdkVersion 28
+        applicationId "jp.studyplus.android.sdk_example_kt"
+        minSdkVersion versions.minSdk
+        targetSdkVersion versions.targetSdk
         versionCode 1
         versionName "1.0"
-
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
-
     }
-
     buildTypes {
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
-
 }
 
 dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-
-    implementation 'com.android.support:appcompat-v7:28.0.0-alpha3'
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+    implementation "com.android.support:appcompat-v7:${versions.supportLibrary}"
     implementation 'com.android.support.constraint:constraint-layout:1.1.2'
 
     implementation project(':studyplus-android-sdk2')
