+++ sdk-example-kt/proguard-rules.pro
