+++ sdk-example-kt/src/androidTest/java/jp/studyplus/android/sdk_example_kt/ExampleInstrumentedTest.kt
@@ -1,0 +1,24 @@
+package jp.studyplus.android.sdk_example_kt
+
+import android.support.test.InstrumentationRegistry
+import android.support.test.runner.AndroidJUnit4
+
+import org.junit.Test
+import org.junit.runner.RunWith
+
+import org.junit.Assert.*
+
+/**
+ * Instrumented test, which will execute on an Android device.
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+@RunWith(AndroidJUnit4::class)
+class ExampleInstrumentedTest {
+    @Test
+    fun useAppContext() {
+        // Context of the app under test.
+        val appContext = InstrumentationRegistry.getTargetContext()
+        assertEquals("jp.studyplus.android.sdk", appContext.packageName)
+    }
+}
