+++ sdk-example-kt/src/main/res/drawable-v24/ic_launcher_foreground.xml
