+++ sdk-example-kt/src/main/res/drawable/ic_launcher_background.xml
