+++ sdk-example-kt/src/main/res/layout/activity_main.xml
