+++ sdk-example-kt/src/main/res/mipmap-mdpi/ic_launcher.png
