+++ sdk-example-kt/src/main/res/mipmap-xhdpi/ic_launcher.png
