+++ sdk-example-kt/src/main/res/mipmap-xxxhdpi/ic_launcher.png
