+++ sdk-example-kt/src/main/res/values/colors.xml
