+++ sdk-example-kt/src/main/res/values/styles.xml
