--- sdk-example/src/androidTest/java/jp/studyplus/android/sdk_example/ExampleInstrumentedTest.kt
@@ -1,24 +1,0 @@
-package jp.studyplus.android.sdk_example
-
-import android.support.test.InstrumentationRegistry
-import android.support.test.runner.AndroidJUnit4
-
-import org.junit.Test
-import org.junit.runner.RunWith
-
-import org.junit.Assert.*
-
-/**
- * Instrumented test, which will execute on an Android device.
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-@RunWith(AndroidJUnit4::class)
-class ExampleInstrumentedTest {
-    @Test
-    fun useAppContext() {
-        // Context of the app under test.
-        val appContext = InstrumentationRegistry.getTargetContext()
-        assertEquals("jp.studyplus.android.sdk", appContext.packageName)
-    }
-}
